Hashtag modules updated
